#471 feat: assemble systemId to Statistic without hard code

+++ src/main/kotlin/com/thoughtworks/archgard/scanner/domain/scanner/statistic/MethodStatistic.kt
@@ -2,7 +2,7 @@ package com.thoughtworks.archgard.scanner.domain.scanner.statistic
 
 data class MethodStatistic(
         val id: String,
-        val systemId: Int,
+        val systemId: Long,
         val moduleName: String? = null,
         val packageName: String,
         val typeName: String,
